+++ prestaspot/includes/class-presta-spot-shortcode.php
@@ -36,6 +36,9 @@ class Presta_Spot_Shortcode
             'show_image' => '',
             'show_name' => '',
             'show_description' => '',
+            'link_text' => '',
+            'link_style' => '',
+            'button_color' => '',
         ), $attributes, 'prestaspot');
 
         $render_args = array(
@@ -44,6 +47,9 @@ class Presta_Spot_Shortcode
             'columns' => (int)$attributes['columns'],
             'layout' => $attributes['layout'],
             'view_mode' => $attributes['view_mode'],
+            'link_text' => $attributes['link_text'],
+            'link_style' => $attributes['link_style'],
+            'button_color' => $attributes['button_color'],
         );
 
         foreach (array('show_image', 'show_name', 'show_description') as $flag) {